docs(examples): move minimal example to examples folder
