+++ examples/apps/minimal/README.md
@@ -1,0 +1,9 @@
+# Minimal demo
+
+This example shows how to create a minimal application.
+
+To run this demo:
+
+```shell
+cargo run -p minimal
+```
