+++ ratatui/Cargo.toml
@@ -158,12 +158,6 @@ required-features = ["crossterm"]
 doc-scrape-examples = true
 
 [[example]]
-name = "minimal"
-required-features = ["crossterm"]
-# prefer to show the more featureful examples in the docs
-doc-scrape-examples = false
-
-[[example]]
 name = "modifiers"
 required-features = ["crossterm"]
 # this example is a bit verbose, so we don't want to include it in the docs
